refactor: move stuff ot separate components
